+++ ErsatzTV.Application/Maintenance/Commands/DeleteOrphanedSubtitlesHandler.cs
@@ -5,29 +5,34 @@ using Microsoft.EntityFrameworkCore;
 
 namespace ErsatzTV.Application.Maintenance;
 
-public class DeleteOrphanedSubtitlesHandler : IRequestHandler<DeleteOrphanedSubtitles, Either<BaseError, Unit>>
+public class DeleteOrphanedSubtitlesHandler(IDbContextFactory<TvContext> dbContextFactory)
+    : IRequestHandler<DeleteOrphanedSubtitles, Either<BaseError, Unit>>
 {
-    private readonly IDbContextFactory<TvContext> _dbContextFactory;
-
-    public DeleteOrphanedSubtitlesHandler(IDbContextFactory<TvContext> dbContextFactory) =>
-        _dbContextFactory = dbContextFactory;
-
     public async Task<Either<BaseError, Unit>> Handle(
         DeleteOrphanedSubtitles request,
         CancellationToken cancellationToken)
     {
         try
         {
-            await using TvContext dbContext = await _dbContextFactory.CreateDbContextAsync(cancellationToken);
+            await using TvContext dbContext = await dbContextFactory.CreateDbContextAsync(cancellationToken);
 
             IEnumerable<int> toDelete = await dbContext.Connection.QueryAsync<int>(
-                @"SELECT S.Id FROM Subtitle S
-                      WHERE S.ArtistMetadataId IS NULL AND S.EpisodeMetadataId IS NULL
-                      AND S.MovieMetadataId IS NULL AND S.MusicVideoMetadataId IS NULL
-                      AND S.OtherVideoMetadataId IS NULL AND S.SeasonMetadataId IS NULL
-                      AND S.ShowMetadataId IS NULL AND S.SongMetadataId IS NULL");
+                """
+                SELECT S.Id FROM Subtitle S
+                    WHERE S.ArtistMetadataId IS NULL AND S.EpisodeMetadataId IS NULL
+                    AND S.ImageMetadataId IS NULL AND S.MovieMetadataId IS NULL
+                    AND S.MusicVideoMetadataId IS NULL AND S.OtherVideoMetadataId IS NULL
+                    AND S.RemoteStreamMetadataId IS NULL AND S.SeasonMetadataId IS NULL
+                    AND S.ShowMetadataId IS NULL AND S.SongMetadataId IS NULL
+                """);
 
-            foreach (int id in toDelete)
+            IEnumerable<int> toDeleteExternal = await dbContext.Connection.QueryAsync<int>(
+                """
+                SELECT S.Id FROM Subtitle S
+                    WHERE S.SubtitleKind = 1 AND (S.Path IS NULL OR S.Path = '')
+                """);
+
+            foreach (int id in toDelete.Concat(toDeleteExternal).Distinct())
             {
                 await dbContext.Connection.ExecuteAsync("DELETE FROM Subtitle WHERE Id = @Id", new { Id = id });
             }